requirements.txt
pymysql==0.10.0
Werkzeug==0.16.1
requests==2.24.0
Flask==1.1.1
bootstrap-flask==1.5.1
PyYAML==5.4.1
